UN-3352 PeriodicUpdater refactor
